Move concerns into TransferBrain namespace
